Move gloo 1.5 to primary cpp, remove gloo 1.3 support
