--- example/cpp-gloo-1.5/runtime-config.json
@@ -1,9 +1,0 @@
-{
-  "type": "envoy_proxy",
-  "abiVersions": ["v0-541b2c1155fffb15ccde92b8324f3e38f7339ba6"],
-  "config": {
-    "rootIds": [
-      "add_header_root_id"
-    ]
-  }
-}
