--- example/cpp-gloo-1.5/toolchain/emar.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-set -euo pipefail
-
-. $(dirname $0)/common.sh
-
-emar "$@"
