+++ example/cpp/BUILD
@@ -18,8 +18,13 @@ wasm_cc_binary(
     srcs = [
         "filter.cc",
     ],
+    additional_linker_inputs = ["@proxy_wasm_cpp_sdk//:jslib"],
+    linkopts = [
+        "--js-library",
+        "external/proxy_wasm_cpp_sdk/proxy_wasm_intrinsics.js",
+    ],
     deps = [
         ":filter_cc_proto",
-        "@envoy_wasm_api//:proxy_wasm_intrinsics",
+        "@proxy_wasm_cpp_sdk//:proxy_wasm_intrinsics",
     ],
 )
