+++ pkg/abi/version.go
@@ -102,11 +102,6 @@ var (
 		Version: Version15x,
 	}
 
-	Version_541b2c1155fffb15ccde92b8324f3e38f7339ba6 = Version{
-		Name:       "v0-541b2c1155fffb15ccde92b8324f3e38f7339ba6",
-		Repository: "https://github.com/yuval-k/envoy-wasm",
-		Commit:     "541b2c1155fffb15ccde92b8324f3e38f7339ba6",
-	}
 	Version_097b7f2e4cc1fb490cc1943d0d633655ac3c522f = Version{
 		Name:       "v0-097b7f2e4cc1fb490cc1943d0d633655ac3c522f",
 		Repository: "https://github.com/envoyproxy/envoy-wasm",
@@ -119,9 +114,6 @@ var (
 	}
 
 	DefaultRegistry = Registry{
-		Version_541b2c1155fffb15ccde92b8324f3e38f7339ba6: {
-			Gloo13,
-		},
 		Version_edc016b1fa5adca3ebd3d7020eaed0ad7b8814ca: {
 			Gloo15,
 		},
